+++ CHANGELOG.md
@@ -13,6 +13,9 @@ TBD
 * **SBBTextarea:** added component
 * **SBBModalView:** renamed to SBBBottomSheet
 * **SBBBottomSheet:** finished component, added isClosable and full-screen mode
+* **SBBDropdown:** added component
+* **SBBListItem:** added textStyle property and made trailingIcon optional
+* **SBBModalView:** use correct title font style
 
 # 0.1.9